distributed_systems/zad4_I3/client/requirements.txt
zeroc-ice==3.6.4
